news_portal/requirements.txt
Django
djangorestframework
psycopg2
drf-yasg
djoser
djangorestframework-simplejwt
django-mptt